pieniä muutoksia

+++ lego/src/data/Data.java
@@ -7,7 +7,7 @@ public class Data {
 	public static boolean shouldRun = true;
 	//public static float distance = 100;
 	public static int direction = 1; 
-	public static int color = 0; //sensorin havaittu väri
+	public static float color = 0; //sensorin havaittu väri
 	public static boolean floodlight = false;  // sensorin valon tila
 	public static int flcolor = 0; //sensorin valon väri
 }
